\name{tensorIA-package}
\alias{tensorIA-package}
\alias{tensorIA}
\docType{package}
\title{
  Integrative analysis based on tensor modelling
}
\description{
  For a grouped multivariates gression model, with or without aparsity assumptions, 
  treating the coefficients as a third-order tensor and borrowing Tucker decomposition to reduce the number of parameters.  
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Xu Liu

Maintainer: Xu Liu <liu.xu@sufe.edu.cn>
}
\references{
  Integrative analysis based on tensor modelling.
}
\keyword{GWAS; Integrative analysis; Tensor estimation; Tucker decomposition.}
